pyproject.toml
[tool.poetry]
name = "djangomusicapp"
version = "0.1.0"
description = ""
authors = ["AbdullahHafez <abdullah.alaa12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.2"
django-model-utils = "^4.2.0"
django-imagekit = "^4.1.0"
Pillow = "^9.2.0"
djangorestframework = "^3.14.0"
Markdown = "^3.4.1"
django-filter = "^22.1"
django-extensions = "^3.2.1"
django-rest-knox = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
